research-paper-analyzer-main/requirements.txt
streamlit
sentence-transformers
pymupdf
groq
pandas
numpy
tqdm
feedparser
requests
python-dotenv